Fix #42
